made changes according to the new working pattern
