+++ app/config/settings/base.py
@@ -50,7 +50,7 @@ DJANGO_APPS = [
 THIRD_PARTY_APPS = ["corsheaders", "django_user_agents"]
 LOCAL_APPS = ["user"]
 
-INSTALLED_APPS = DJANGO_APPS + THIRD_PARTY_APPS + LOCAL_APPS
+INSTALLED_APPS = DJANGO_APPS + THIRD_PARTY_APPS  # + LOCAL_APPS
 
 
 # MIDDLEWARE
