+++ app/event/utils/model.py
@@ -4,7 +4,7 @@ from django.conf import settings
 from django.contrib.auth import get_user_model
 from django.db import models
 from django.db.models.signals import pre_delete, post_delete
-from app.event.utils.thread_local import get_current_request
+from .thread_local import get_current_request
 
 MODEL_REGISTRY = {}
