+++ app/event/views/EventView.py
@@ -1,0 +1,44 @@
+from rest_framework import status
+from rest_framework.decorators import api_view
+from rest_framework.response import Response
+
+from ..models.event import Event
+from ..serializers.eventSerializer import EventSerializer
+
+
+@api_view(['GET', 'POST'])
+def event_list(request):
+
+    if request.method == 'GET':
+        event = Event.objects.all()
+        serializer = EventSerializer(event, many = True)
+        return Response(serializer.data)
+
+    elif request.method == 'POST':
+        serializer = EventSerializer(data=request.data, many=True)
+        if serializer.is_valid():
+            serializer.save()
+            return Response(serializer.data, status=status.HTTP_201_CREATED)
+        return Response(serializer.errors, status.HTTP_400_BAD_REQUEST)
+
+
+@api_view(['GET', 'PUT', 'DELETE'])
+def event_detail(request, event_id):
+    try :
+        event = Event.objects.get(pk= event_id)
+    except Event.DoesNotExist :
+        return Response(status=status.HTTP_404_NOT_FOUND)
+
+    if request.method == 'GET' :
+        serializer = EventSerializer(event)
+        return Response(serializer.data, status=status.HTTP_200_OK)
+    elif request.method == 'PUT':
+        serializer = EventSerializer(event, data=request.data)
+        if serializer.is_valid():
+            serializer.save()
+            return Response(serializer.data)
+        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+
+    elif request.method == 'DELETE':
+        event.delete()
+        return Response(status=status.HTTP_204_NO_CONTENT)
